Fix `-isystem` includes being incorrectly passed to execute_process()
The external command was being invoked with arguments: ["clang", ...,
"-isystem  /foo/bar/baz", ...], which meant that the compiler treated
the whitespace between the flag and the path as being part of the path
itself. (i.e. it was looking for "  /foo/bar/baz").

Fixed by passing the argument to execute_process() as
'-isystem/foo/bar/baz'; I also needed to ensure we normalized the value
of CMAKE_INCLUDE_SYSTEM_FLAG_<LANG>, which itself contains trailing
whitespace in some cases.