Copy IMPORT_PREFIX and IMPORT_SUFFIX target properties for unity targets.
For unity targets, the PREFIX and SUFFIX target properties were being
copied so that library output file names would make the non-unity
target.  However, this wasn't the case for IMPORT_PREFIX and
IMPORT_SUFFIX which control the import library file name on Windows.  As
a result, the unity target and the original target differed in the
output file name for import libraries.

This fix simply adds IMPORT_PREFIX and IMPORT_SUFFIX to the list of
properties being copied from the original target.